+++ src/linters/spacing.js
@@ -1,17 +1,9 @@
 /* eslint-disable import/no-commonjs */
-const fs = require('fs');
-const path = require('path');
 const stylelint = require('stylelint');
 
-const spacing = fs.readFileSync(path.join(__dirname, '..', 'styles', 'spacing.css'), 'UTF-8');
-const cssVarRegex = /--[a-z0-9-]+/g;
-const validSpacingSettings = spacing.match(cssVarRegex);
-
-const cssValueRegex = /[0-9]+px/g;
-const valueKeys = spacing.match(cssValueRegex);
-
 const ruleName = 'mds/spacing';
-const url = 'https://mavenlink.github.io/design-system/master/#/Brand%20Identity/Spacing';
+const url =
+  'https://mavenlink.github.io/design-system/master/#/Brand%20Identity/Spacing';
 
 // Note this can be replaced by native String.prototype.matchAll when node is upgraded past: 12.0.0
 const matchAll = (string, regex) => {
@@ -33,7 +25,10 @@ const matchAll = (string, regex) => {
         matchValue.index = index + buffer;
         matchValue[0] = match;
         finalMatches.push(matchValue);
-        splicedString = splicedString.substring(matchValue.index + match.length, splicedString.length);
+        splicedString = splicedString.substring(
+          matchValue.index + match.length,
+          splicedString.length,
+        );
         buffer += index + match.length;
         index = splicedString.indexOf(match);
       }
@@ -43,10 +38,18 @@ const matchAll = (string, regex) => {
   return finalMatches;
 };
 
-function zip(obj, value, index) {
-  return { ...obj, [value]: validSpacingSettings[index] };
-}
-const fixmap = valueKeys.reduce(zip, {});
+const fixmap = {
+  '0.125rem': '--spacing-x-small',
+  '2px': '--spacing-x-small',
+  '0.25rem': '--spacing-small',
+  '4px': '--spacing-small',
+  '0.5rem': '--spacing-medium',
+  '8px': '--spacing-medium',
+  '1rem': '--spacing-large',
+  '16px': '--spacing-large',
+  '2rem': '--spacing-x-large',
+  '32px': '--spacing-x-large',
+};
 
 const messages = stylelint.utils.ruleMessages(ruleName, {
   rejected: `To promote design consistency: use MDS spacing variables for margin and padding.
@@ -70,83 +73,104 @@ const properties = [
 
 const fixes = [];
 
-module.exports = stylelint.createPlugin(ruleName, (primary, secondary, context) => {
-  return (root, result) => {
-    root.walkDecls((declaration) => {
-      const property = declaration.prop;
+module.exports = stylelint.createPlugin(
+  ruleName,
+  (primary, secondary, context) => {
+    return (root, result) => {
+      root.walkDecls((declaration) => {
+        const property = declaration.prop;
 
-      if (properties.includes(property)) {
-        const currentValue = declaration.value;
+        if (properties.includes(property)) {
+          const currentValue = declaration.value;
 
-        const invalidSpacingValue = Object.keys(fixmap).some((spacingVariable) => {
-          const pixelValueMatches = [...matchAll(currentValue, /-?\d+px/g)];
-          return pixelValueMatches.some(pixValueMatch => pixValueMatch[0] === spacingVariable);
-        });
+          const invalidSpacingValue = Object.keys(fixmap).some(
+            (spacingVariable) => {
+              const pixelValueMatches = [...matchAll(currentValue, /-?[\d.]+(px|rem)/g)];
+              return pixelValueMatches.some(
+                pixValueMatch => pixValueMatch[0] === spacingVariable,
+              );
+            },
+          );
 
-        if (invalidSpacingValue) {
-          if (context.fix) {
-            let valueToFix = currentValue;
-            Object.keys(fixmap).forEach((spacingVariable) => {
-              const pixelValueMatches = [...matchAll(valueToFix, (/-?\d+px/g))];
-              pixelValueMatches.reverse().forEach((pixelValueMatch) => {
-                const matchIsNegative = pixelValueMatch[0][0] === '-';
+          if (invalidSpacingValue) {
+            if (context.fix) {
+              let valueToFix = currentValue;
+              Object.keys(fixmap).forEach((spacingVariable) => {
+                const pixelValueMatches = [...matchAll(valueToFix, /-?[\d.]+(px|rem)/g)];
+                pixelValueMatches.reverse().forEach((pixelValueMatch) => {
+                  const matchIsNegative = pixelValueMatch[0][0] === '-';
 
-                let matchingString;
-                let mdsVariable;
-                let secondHalf;
-                if (matchIsNegative) {
-                  matchingString = pixelValueMatch[0].substring(1, pixelValueMatch[0].length);
-                  mdsVariable = `calc(-1 * var(${fixmap[spacingVariable]}))`;
-                  secondHalf = valueToFix.substring(pixelValueMatch.index + 1
-                    + spacingVariable.length, valueToFix.length);
-                } else {
-                  matchingString = pixelValueMatch[0];
-                  mdsVariable = `var(${fixmap[spacingVariable]})`;
-                  secondHalf = valueToFix.substring(pixelValueMatch.index
-                    + spacingVariable.length, valueToFix.length);
-                }
+                  let matchingString;
+                  let mdsVariable;
+                  let secondHalf;
+                  if (matchIsNegative) {
+                    matchingString = pixelValueMatch[0].substring(
+                      1,
+                      pixelValueMatch[0].length,
+                    );
+                    mdsVariable = `calc(-1 * var(${fixmap[spacingVariable]}))`;
+                    secondHalf = valueToFix.substring(
+                      pixelValueMatch.index + 1 + spacingVariable.length,
+                      valueToFix.length,
+                    );
+                  } else {
+                    matchingString = pixelValueMatch[0];
+                    mdsVariable = `var(${fixmap[spacingVariable]})`;
+                    secondHalf = valueToFix.substring(
+                      pixelValueMatch.index + spacingVariable.length,
+                      valueToFix.length,
+                    );
+                  }
 
-                if (matchingString === spacingVariable) {
-                  const firstHalf = valueToFix.substring(0, pixelValueMatch.index);
-                  valueToFix = `${firstHalf}${mdsVariable}${secondHalf}`;
-                }
+                  if (matchingString === spacingVariable) {
+                    const firstHalf = valueToFix.substring(
+                      0,
+                      pixelValueMatch.index,
+                    );
+                    valueToFix = `${firstHalf}${mdsVariable}${secondHalf}`;
+                  }
+                });
               });
-            });
-            fixes.push([declaration, valueToFix]);
+              fixes.push([declaration, valueToFix]);
 
-            let shouldImport = true;
-            try {
-              require.resolve('@mavenlink/design-system');
-            } catch (e) {
-              shouldImport = false;
-            }
+              let shouldImport = true;
+              try {
+                require.resolve('@mavenlink/design-system');
+              } catch (e) {
+                shouldImport = false;
+              }
 
-            const importString = '@import \'@mavenlink/design-system/src/styles/spacing.css\';';
-            if (!root.toString().includes(importString) && shouldImport) {
-              root.prepend(importString);
+              const importString =
+                "@import '@mavenlink/design-system/src/styles/spacing.css';";
+              if (!root.toString().includes(importString) && shouldImport) {
+                root.prepend(importString);
+              }
+            } else {
+              const violation = { ruleName, node: declaration, result };
+              stylelint.utils.report({
+                ...violation,
+                message: messages.rejected,
+              });
             }
-          } else {
-            const violation = { ruleName, node: declaration, result };
-            stylelint.utils.report({ ...violation, message: messages.rejected });
           }
         }
-      }
-    });
+      });
 
-    // eslint-disable-next-line prefer-arrow-callback
-    return new Promise(function (resolve) {
       // eslint-disable-next-line prefer-arrow-callback
-      setTimeout(function () {
+      return new Promise(function (resolve) {
         // eslint-disable-next-line prefer-arrow-callback
-        fixes.forEach(function (fixTuple) {
-          // eslint-disable-next-line no-param-reassign
-          fixTuple[0].value = fixTuple[1];
-        });
-        resolve();
-      }, 1);
-    });
-  };
-});
+        setTimeout(function () {
+          // eslint-disable-next-line prefer-arrow-callback
+          fixes.forEach(function (fixTuple) {
+            // eslint-disable-next-line no-param-reassign
+            fixTuple[0].value = fixTuple[1];
+          });
+          resolve();
+        }, 1);
+      });
+    };
+  },
+);
 
 module.exports.ruleName = ruleName;
 module.exports.targetedProperties = properties;