+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
 - Append new items to make git merging easier.
+- feature(rems): converts size and spacing css variable to rems for better accessibility support
 
 ## 0.90.0
 - feat(select-cell): allow prefilled values